// game/Custom/ShanghaiNY/src/interface.d.ts
export interface ICreateParams {

    /** CreateParams playersPerGroup */
    playersPerGroup?: (number|null);

    /** CreateParams rounds */
    rounds?: (number|null);

    /** CreateParams gameType */
    gameType?: (number|null);

    /** CreateParams version */
    version?: (number|null);

    /** CreateParams a */
    a?: (number|null);

    /** CreateParams b */
    b?: (number|null);

    /** CreateParams c */
    c?: (number|null);

    /** CreateParams d */
    d?: (number|null);

    /** CreateParams eH */
    eH?: (number|null);

    /** CreateParams eL */
    eL?: (number|null);

    /** CreateParams s */
    s?: (number|null);

    /** CreateParams p */
    p?: (number|null);

    /** CreateParams b0 */
    b0?: (number|null);

    /** CreateParams b1 */
    b1?: (number|null);

    /** CreateParams participationFee */
    participationFee?: (number|null);

    /** CreateParams nextPhaseKey */
    nextPhaseKey?: (string|null);

    /** CreateParams id */
    id?: (string|null);
}
// ...
export interface IGameState {

    /** GameState groups */
    groups?: (GameState.IGroup[]|null);
}

/** Represents a GameState. */
export class GameState implements IGameState {

    /**
     * Constructs a new GameState.
     * @param [properties] Properties to set
     */
    constructor(properties?: IGameState);

    /** GameState groups. */
    public groups: GameState.IGroup[];

    /**
     * Creates a new GameState instance using the specified properties.
     * @param [properties] Properties to set
     * @returns GameState instance
     */
    public static create(properties?: IGameState): GameState;

    /**
     * Encodes the specified GameState message. Does not implicitly {@link GameState.verify|verify} messages.
     * @param message GameState message or plain object to encode
     * @param [writer] Writer to encode to
     * @returns Writer
     */
    public static encode(message: IGameState, writer?: $protobuf.Writer): $protobuf.Writer;

    /**
     * Encodes the specified GameState message, length delimited. Does not implicitly {@link GameState.verify|verify} messages.
     * @param message GameState message or plain object to encode
     * @param [writer] Writer to encode to
     * @returns Writer
     */
    public static encodeDelimited(message: IGameState, writer?: $protobuf.Writer): $protobuf.Writer;

    /**
     * Decodes a GameState message from the specified reader or buffer.
     * @param reader Reader or buffer to decode from
     * @param [length] Message length if known beforehand
     * @returns GameState
     * @throws {Error} If the payload is not a reader or valid buffer
     * @throws {$protobuf.util.ProtocolError} If required fields are missing
     */
    public static decode(reader: ($protobuf.Reader|Uint8Array), length?: number): GameState;

    /**
     * Decodes a GameState message from the specified reader or buffer, length delimited.
     * @param reader Reader or buffer to decode from
     * @returns GameState
     * @throws {Error} If the payload is not a reader or valid buffer
     * @throws {$protobuf.util.ProtocolError} If required fields are missing
     */
    public static decodeDelimited(reader: ($protobuf.Reader|Uint8Array)): GameState;

    /**
     * Verifies a GameState message.
     * @param message Plain object to verify
     * @returns `null` if valid, otherwise the reason why it is not
     */
    public static verify(message: { [k: string]: any }): (string|null);

    /**
     * Creates a GameState message from a plain object. Also converts values to their respective internal types.
     * @param object Plain object
     * @returns GameState
     */
    public static fromObject(object: { [k: string]: any }): GameState;

    /**
     * Creates a plain object from a GameState message. Also converts values to other types if specified.
     * @param message GameState
     * @param [options] Conversion options
     * @returns Plain object
     */
    public static toObject(message: GameState, options?: $protobuf.IConversionOptions): { [k: string]: any };

    /**
     * Converts this GameState to JSON.
     * @returns JSON object
     */
    public toJSON(): { [k: string]: any };
}

export namespace GameState {

    /** Properties of a Group. */
    interface IGroup {

        /** Group playerNum */
        playerNum?: (number|null);

        /** Group probs */
        probs?: (boolean[]|null);

        /** Group ones */
        ones?: (boolean[]|null);

        /** Group mins */
        mins?: (number[]|null);
    }

    /** Represents a Group. */
    class Group implements IGroup {

        /**
         * Constructs a new Group.
         * @param [properties] Properties to set
         */
        constructor(properties?: GameState.IGroup);

        /** Group playerNum. */
        public playerNum: number;

        /** Group probs. */
        public probs: boolean[];

        /** Group ones. */
        public ones: boolean[];

        /** Group mins. */
        public mins: number[];

        /**
         * Creates a new Group instance using the specified properties.
         * @param [properties] Properties to set
         * @returns Group instance
         */
        public static create(properties?: GameState.IGroup): GameState.Group;

        /**
         * Encodes the specified Group message. Does not implicitly {@link GameState.Group.verify|verify} messages.
         * @param message Group message or plain object to encode
         * @param [writer] Writer to encode to
         * @returns Writer
         */
        public static encode(message: GameState.IGroup, writer?: $protobuf.Writer): $protobuf.Writer;

        /**
         * Encodes the specified Group message, length delimited. Does not implicitly {@link GameState.Group.verify|verify} messages.
         * @param message Group message or plain object to encode
         * @param [writer] Writer to encode to
         * @returns Writer
         */
        public static encodeDelimited(message: GameState.IGroup, writer?: $protobuf.Writer): $protobuf.Writer;

        /**
         * Decodes a Group message from the specified reader or buffer.
         * @param reader Reader or buffer to decode from
         * @param [length] Message length if known beforehand
         * @returns Group
         * @throws {Error} If the payload is not a reader or valid buffer
         * @throws {$protobuf.util.ProtocolError} If required fields are missing
         */
        public static decode(reader: ($protobuf.Reader|Uint8Array), length?: number): GameState.Group;

        /**
         * Decodes a Group message from the specified reader or buffer, length delimited.
         * @param reader Reader or buffer to decode from
         * @returns Group
         * @throws {Error} If the payload is not a reader or valid buffer
         * @throws {$protobuf.util.ProtocolError} If required fields are missing
         */
        public static decodeDelimited(reader: ($protobuf.Reader|Uint8Array)): GameState.Group;

        /**
         * Verifies a Group message.
         * @param message Plain object to verify
         * @returns `null` if valid, otherwise the reason why it is not
         */
        public static verify(message: { [k: string]: any }): (string|null);

        /**
         * Creates a Group message from a plain object. Also converts values to their respective internal types.
         * @param object Plain object
         * @returns Group
         */
        public static fromObject(object: { [k: string]: any }): GameState.Group;

        /**
         * Creates a plain object from a Group message. Also converts values to other types if specified.
         * @param message Group
         * @param [options] Conversion options
         * @returns Plain object
         */
        public static toObject(message: GameState.Group, options?: $protobuf.IConversionOptions): { [k: string]: any };

        /**
         * Converts this Group to JSON.
         * @returns JSON object
         */
        public toJSON(): { [k: string]: any };
    }
}

/** Properties of a PlayerState. */
export interface IPlayerState {

    /** PlayerState groupIndex */
    groupIndex?: (number|null);

    /** PlayerState positionIndex */
    positionIndex?: (number|null);

    /** PlayerState seatNumber */
    seatNumber?: (string|null);

    /** PlayerState stage */
    stage?: (number|null);

    /** PlayerState stageIndex */
    stageIndex?: (number|null);

    /** PlayerState choices */
    choices?: (PlayerState.IChoice[]|null);

    /** PlayerState profits */
    profits?: (number[]|null);

    /** PlayerState finalProfit */
    finalProfit?: (number|null);

    /** PlayerState surveyAnswers */
    surveyAnswers?: (string[]|null);

    /** PlayerState roundIndex */
    roundIndex?: (number|null);

    /** PlayerState mobile */
    mobile?: (string|null);
}
// ...
    interface IChoice {

        /** Choice c1 */
        c1?: (number|null);

        /** Choice c2 */
        c2?: (number[]|null);

        /** Choice c */
        c?: (number|null);
    }
// ...
export interface IMoveParams {

    /** MoveParams seatNumber */
    seatNumber?: (string|null);

    /** MoveParams c1 */
    c1?: (number|null);

    /** MoveParams c2 */
    c2?: (number[]|null);

    /** MoveParams surveys */
    surveys?: (string[]|null);

    /** MoveParams nextRoundIndex */
    nextRoundIndex?: (number|null);
}
// ...
export interface IPushParams {
}